streamlit_app/requirements.txt
beautifulsoup4==4.11.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
pandas==1.5.2
protobuf==3.20.3
PyYAML==6.0
pymongo==4.3.3
streamlit==1.16.0
streamlit_ace==0.1.1
streamlit_agraph==0.0.42
streamlit_nested_layout==0.1.1
streamlit_toggle_switch==1.0.2
elasticsearch==8.6.0
st-clickable-images==0.0.3
torch==1.13.1
transformers==4.25.1